restructured a bit
